+++ premake5.lua
@@ -333,6 +333,7 @@ group "exts"
     -- Linux Only
     if os.target() == "linux" then
         include ("source/extensions/omni.isaac.cortex")
+        include ("source/extensions/omni.isaac.cortex_sync")
         -- include ("source/extensions/omni.isaac.robot_engine_bridge")
         include ("source/extensions/omni.isaac.gxf_bridge")
         include ("source/extensions/omni.isaac.ros_bridge")